+++ src/util/applyFeatureTo.js
@@ -1,27 +1,25 @@
-import { always, compose, cond, curry, equals, T } from 'ramda';
+import { curry } from 'ramda';
 import { isFunction } from 'ramda-adjunct';
 import { getHandler, setFeatureIsLoaded } from '../lens/feature';
-import { ensurePromise } from './async';
+import { ensurePromise } from './promise';
 import pipeAsync from './pipeAsync';
 import { HANDLER_NOT_READY_RESULT } from '../constants';
 
-const mapRightResult = setFeatureIsLoaded(true);
+const transformRightResult = setFeatureIsLoaded(true);
 
-const resolveHandler = curry((app, handler) => ensurePromise(handler(app)));
+const applyFeatureTo = curry(async (app, feature) => {
+  const handler = getHandler(feature);
+  const result = await ensurePromise(handler(app));
 
-const callFeatureWith = curry((app, feature) =>
-  compose(resolveHandler(app), getHandler)(feature),
-);
+  if (result === HANDLER_NOT_READY_RESULT) {
+    return feature;
+  }
 
-const applyFeatureTo = curry((app, feature) =>
-  pipeAsync(
-    callFeatureWith,
-    cond([
-      [equals(HANDLER_NOT_READY_RESULT), always(feature)],
-      [isFunction, result => pipeAsync(mapRightResult, result)(feature)],
-      [T, compose(mapRightResult, always(feature))],
-    ]),
-  )(app, feature),
-);
+  if (isFunction(result)) {
+    return pipeAsync(transformRightResult, result)(feature);
+  }
+
+  return transformRightResult(feature);
+});
 
 export default applyFeatureTo;
