+++ src/util/debug.js
@@ -1,5 +1,5 @@
 import debug from 'debug';
-import { apply, bind, compose, tap, unapply } from 'ramda';
+import { apply, bind, pipe, tap, unapply } from 'ramda';
 import { PKG_NAME } from '../constants';
 
 const baseDebug = debug(PKG_NAME);
@@ -8,7 +8,7 @@ const debugIt = baseDebug;
 
 const extend = bind(debugIt.extend, debugIt);
 
-const createDebug = compose(unapply, apply, extend);
+const createDebug = pipe(extend, apply, unapply);
 
 const debugItFp = tap(createDebug('fp'));
